Reorganize class structure, 'HashStoreInterface' is now 'HashStore' and 'FileHashStore' implements 'HashStore'
